Update for clarity and consistency
Previous verbiage sounded like top and left were offsets, like it'd scroll that amount of pixels from the current scroll location, which isn't the case (at least in Chrome).
The examples specifying the same coordinates to be clearer.
Would like to add in the `options` docs that each index is optional. Added a suggestion, but I'm not sure how that's usually noted in the docs.